pyproject.toml
[tool.poetry]
name = "file-convertion-tools"
version = "0.0.14"
description = "common file conversions"
authors = ["Gustav Rasmussen"]
readme = "README.md"
packages = [{include = "file_convertion_tools"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
pytest = "^8.3.3"
pypdf = "^5.1.0"
pyyaml = "^6.0.2"
tabula-py = "^2.10.0"
tabulate = "^0.9.0"
py7zr = "^0.22.0"
moviepy = "^1.0.3"
pytest-cov = "^6.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ruff = "^0.7.2"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
